chore: Create new component and move logic from ProductPage
